pyproject.toml
[tool.poetry]
name = "coffeegis-geoprocessor-proto"
version = "0.1.0"
description = ""
authors = ["smbrine <smbrine@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.6.3"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.28"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
python-dotenv = "^1.0.1"
redis = "^5.0.2"
elasticsearch = {extras = ["async"], version = "^8.12.1"}
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
pylint = "^3.1.0"
flake8 = "^7.0.0"
ruff = "^0.3.2"

[tool.poetry.group.migration.dependencies]
alembic = "^1.13.1"

[tool.poetry.group.grpc-deploy.dependencies]
grpcio = "^1.62.0"
grpcio-tools = "^1.62.0"
grpcio-reflection = "^1.62.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
